Refactor todos router to enforce user authentication and associate todos with the authenticated user

+++ routers/todos.py
@@ -7,6 +7,7 @@ from fastapi import  Depends, HTTPException, Path
 from database import SessionLocal 
 from models import Todos
 from starlette import status
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -18,6 +19,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -26,28 +28,36 @@ class TodoRequest(BaseModel):
     complete: bool
 
 @router.get("/")
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency,db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    return db.query(Todos).filter(Todos.owner_id== user.get('id')).all()
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_by_id(db:db_dependency,todo_id:int = Path(gt=0)):
-    todo_model= db.query(Todos).filter(Todos.id== todo_id).first()
+async def read_by_id(user: user_dependency, db:db_dependency,todo_id:int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    todo_model= db.query(Todos).filter(Todos.id== todo_id).filter(Todos.owner_id==user.get('id')).first()
 
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail='Todo not found.')
 
 @router.post("/todo",status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request:TodoRequest):
-    todo_model= Todos(**todo_request.dict())
+async def create_todo(user: user_dependency, db: db_dependency, todo_request:TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    todo_model= Todos(**todo_request.model_dump(),owner_id=user.get('id'))
 
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency,todo_request:TodoRequest, todo_id:int=Path(gt=0) ):
-
-    todo_model=db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user: user_dependency,db: db_dependency,todo_request:TodoRequest, todo_id:int=Path(gt=0) ):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    
+    todo_model=db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found.')
     
@@ -60,12 +70,15 @@ async def update_todo(db: db_dependency,todo_request:TodoRequest, todo_id:int=Pa
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id:int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency,db: db_dependency, todo_id:int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication failed')
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id==user.get('id')).first()
 
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found.')
-    db.query(Todos).filter(Todos.id==todo_id).delete()
+    db.query(Todos).filter(Todos.id==todo_id).filter(Todos.owner_id==user.get('id')).delete()
 
     db.commit()
